+++ src/applications/financial-status-report/components/otherAssets/EnhancedVehicleRecord.jsx
